尽量统一Swoole作为Server时，各种工具可以通用，如router，验证器等
