+++ src/dispatcher/Command.php
@@ -20,7 +20,6 @@ use nb\Debug;
 use nb\Router;
 use nb\Pool;
 
-
 /**
  * Command
  *
@@ -30,10 +29,14 @@ use nb\Pool;
  * @author: collin <collin@nb.cx>
  * @date: 2017/12/3
  */
-class Command extends Driver  implements ICommand {
+class Command extends Driver implements ICommand {
 
-    //请求控制器方法的参数
-    protected $params = [];
+    protected $pathinfo;
+
+    /**
+     * @var Input
+     */
+    protected $in;
 
     public function configure(Pack $cmd) {
         $cmd->setName('action')
@@ -45,8 +48,36 @@ class Command extends Driver  implements ICommand {
 
     public function execute(Input $input, Output $output) {
         Debug::start();
-        $this->params = $input->getArgument('args');
-        $this->with(trim($input->getArgument('router')));
+        $this->in = $input;
+        $this->pathinfo = trim($input->getArgument('router'));
+        $this->run();
+    }
+
+    protected function input(\ReflectionClass $controller, $app) {
+        if($this->in) {
+            return $this->in->getArgument('args');
+        }
+        return [];
+    }
+
+    public function run() {
+        $router = Router::driver();
+        $this->pathinfo and $router->pathinfo = $this->pathinfo;
+        $router = $router->mustAnalyse();
+
+        //如果访问的模块，加载模块配置
+        $router->module and $this->module($router->module);
+
+        ///如果加载不成功，作为404处理
+        $class = $router->class;
+        if(!$class) {
+            return Pool::object('nb\event\Framework')->notfound();
+        }
+        //过滤掉禁止访问的方法
+        if (in_array($router->function,Config::$o->notFunc)) {
+            return Pool::object('nb\event\Framework')->notfound();
+        }
+        $this->go($class,$router->function);
     }
 
     function interact(Input $input, Output $output){}
@@ -57,71 +88,5 @@ class Command extends Driver  implements ICommand {
      * @param Output $output An OutputInterface instance
      */
     function initialize(Input $input, Output $output){}
-
-    /**
-     * 处理控制台指令
-     */
-    public function run() {
-        $this->params = $_SERVER['argv'];
-        array_splice($this->params,0,2);
-        $this->with();
-    }
-
-    protected function with($pathinfo=null) {
-        //路由解析前的回调函数
-        //可以重定路由，可以修改路由配置等
-        Pool::object('nb\\event\\Framework')->redirect();
-
-        $router = Router::ins();
-        $pathinfo and $router->pathinfo = $pathinfo;
-        $router = $router->mustAnalyse();
-
-        //如果访问的模块，加载模块配置
-        if($router->module) {
-            $this->module($router->module);
-        }
-
-        ///如果加载不成功，作为404处理
-        $class = $router->class;
-        if(!$class) {
-            return Pool::object('nb\\event\\Framework')->notfound();
-        }
-        //过滤掉禁止访问的方法
-        if (in_array($router->function,Config::$o->notFunc)) {
-            return Pool::object('nb\\event\\Framework')->notfound();
-        }
-        $this->go($class,$router->function);
-
-    }
-
-    public function go($class, $function) {
-        // TODO: Implement go() method.
-        $class = new \ReflectionClass($class);
-
-        //创建当前控制器对象，并放入池子
-        $app = Pool::value('controller',$class->newInstance());
-
-        $return = null;
-        //判断用户是否构建了__before方法,如果构建，则只有__before为true，才进行处理
-        if (!$class->hasMethod('__before') || $app->__before()) {
-            if ($class->hasMethod($function)) {
-                $method = new \ReflectionMethod($app, $function);
-                if (!$method->isPublic() || $method->isStatic()) {
-                    Pool::object('nb\\event\\Framework')->notfound();
-                    return;
-                }
-
-                $return = call_user_func_array([$app,$function],$this->params);
-            }
-            else {
-                return Pool::object('nb\\event\\Framework')->notfound();;
-            }
-        }
-
-        //判断用户是否构建了__after方法,如果构建，则执行
-        if ($class->hasMethod('__after')) {
-            $app->__after($return);
-        }
-    }
 
 }
