+++ src/request/Http.php
@@ -21,7 +21,7 @@ use nb\Pool;
  * @author: collin <collin@nb.cx>
  * @date: 2017/11/28
  */
-class Http extends Driver {
+class Http extends Php {
 
     /**
      * @var \swoole\http\Request
@@ -33,6 +33,7 @@ class Http extends Driver {
         if(!$this->req) {
             $this->req = new \StdClass();
         }
+        Pool::object('nb\event\Framework')->request($this);
     }
 
     public function _input() {
