+++ src/request/Php.php
@@ -9,6 +9,8 @@
  */
 namespace nb\request;
 
+use nb\Pool;
+
 /**
  * Native
  *
@@ -20,7 +22,6 @@ namespace nb\request;
  */
 class Php extends Driver {
 
-    // php://input
     public function __construct() {
         $this->tmp = [
             'input'   => file_get_contents('php://input'),
@@ -31,6 +32,7 @@ class Php extends Driver {
             'cookie'  => &$_COOKIE,
             'server'  => &$_SERVER
         ];
+        Pool::object('nb\event\Framework')->request($this);
     }
 
     /**
@@ -75,15 +77,15 @@ class Php extends Driver {
         return false;
     }
 
-    public function _requestTime() {
+    protected function _requestTime() {
         return $this->server['REQUEST_TIME'];
     }
 
-    public function _requestTimeFloat() {
+    protected function _requestTimeFloat() {
         return $this->server['REQUEST_TIME_FLOAT'];
     }
 
-    public function _method(){
+    protected function _method(){
         return $_SERVER['REQUEST_METHOD'];
     }
 
@@ -94,7 +96,7 @@ class Php extends Driver {
      * @access public
      * @return boolean
      */
-    public function _isMobile() {
+    protected function _isMobile() {
         $userAgent = $this->_agent();
         return preg_match('/android.+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|symbian|treo|up\.(browser|link)|vodafone|wap|windows (ce|phone)|xda|xiino/i', $userAgent) || preg_match('/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(di|rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i', substr($userAgent, 0, 4));
     }
@@ -105,7 +107,7 @@ class Php extends Driver {
      * @param string $domain 域名
      * @return string
      */
-    public function _domain() {
+    protected function _domain() {
         return $this->scheme . '://' . $this->host;
     }
 
@@ -115,7 +117,7 @@ class Php extends Driver {
      * @access public
      * @return string
      */
-    public function _agent() {
+    protected function _agent() {
         return $_SERVER['HTTP_USER_AGENT'];
     }
 
@@ -125,7 +127,7 @@ class Php extends Driver {
      * @param boolean $adv 是否进行高级模式获取（有可能被伪装）
      * @return mixed
      */
-    public function _ip() {
+    protected function _ip() {
         if (isset($_SERVER['HTTP_X_FORWARDED_FOR'])) {
             $arr = explode(',', $_SERVER['HTTP_X_FORWARDED_FOR']);
             $pos = array_search('unknown', $arr);
@@ -321,6 +323,67 @@ class Php extends Driver {
         }
 
         return '/' . ltrim(urldecode($pathInfo), '/');
+    }
+
+    /**
+     * 获取url的扩展名
+     * @return null
+     */
+    protected function _ext() {
+        $url = $this->uri;
+        $urlinfo =  parse_url($url);
+        $file = basename($urlinfo['path']);
+        if(strpos($file,'.') !== false) {
+            $ext = explode('.',$file);
+            return $ext[count($ext)-1];
+        }
+        return null;
+    }
+
+
+    /**
+     * 获取表单数据，返回一个结果数组
+     * @param string $method
+     * @param null $args
+     * @return array
+     */
+    public function form($method='request',array $args=null) {
+        $method = $method === 'auto'?strtolower($this->method()):$method;
+        $input = [];
+        switch ($method) {
+            case 'request':
+                $input = $this->request;
+                break;
+            case 'post':
+                $input = $this->post;
+                break;
+            case 'get':
+                $input = $this->get;
+                break;
+            case 'request':
+                $input = $this->request;
+                break;
+            case 'input':
+                $input = $this->input;
+                break;
+            case 'put':
+                parse_str($this->input, $input);
+                break;
+            case 'files':
+                $input = $this->files;
+                break;
+            case 'server':
+                $input = $this->server;
+                break;
+        }
+        if($args) {
+            $_input = [];
+            foreach ($args as $arg) {
+                $_input[$arg] = isset($input[$arg])?$input[$arg]:null;
+            }
+            $input = $_input;
+        }
+        return $input;
     }
 
 }
